+++ CHANGELOG.md
@@ -85,6 +85,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Log cpio errors more prominently. #1615
 - Improved syncuser conflict help text. #1614
 - Parallelized overlay build. #1018
+- Parallelized and optimized overlay build. #1018
 
 ### Removed